+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+- [#800]: `defmt-macros`: Fix generic trait bounds in Format derive macro
+
+[#800]: https://github.com/knurling-rs/defmt/pull/800
+
 ## [v0.3.6] - 2024-02-05
 
 - [#804]: `CI`: Remove mdbook strategy